Refactor folders structure to have src/ containing frontend, middleware and database, and tests/ directory containing unit-tests; Ignore dist/ folder, because it will be created by CI/CD;
