+++ index.html
@@ -6,7 +6,7 @@
 
     <title>Fake PEV Shopping</title>
 
-    <link rel="shortcut icon" type="image/png" href="/assets/icons/icons8-engine-48.png">
+    <link rel="shortcut icon" type="image/png" href="/src/frontend/assets/icons/icons8-engine-48.png">
     <link rel="stylesheet" type="text/css" href="/dist/main.css">
 </head>
 <body>
