+++ package.json
@@ -17,14 +17,14 @@
   "bugs": {
     "url": "https://github.com/JSHoliq/Fake-PEV-Shopping/issues"
   },
-  "main": "src/app.js",
+  "main": "src/frontend/app.js",
   "scripts": {
     "test": "echo \"Tests not ready yet!\"",
-    "prettier": "prettier --write src/ middleware/ database/ !database/data",
-    "prettier:backend": "prettier --write middleware/ database/ !database/data",
-    "eslint:backend": "eslint middleware/** database/**",
+    "prettier": "prettier --write src/ !src/database/data",
+    "prettier:backend": "prettier --write src/middleware/ src/database/ !src/database/data",
+    "eslint:backend": "eslint src/middleware/** src/database/**",
     "start": "set \"NODE_ONLY=false\" && webpack-dev-server",
-    "serve": "set \"NODE_ONLY=true\" && nodemon --exec \"npm run eslint:backend && npm run prettier:backend && node middleware/index.js\"",
+    "serve": "set \"NODE_ONLY=true\" && nodemon --exec \"npm run eslint:backend && npm run prettier:backend && node src/middleware/index.js\"",
     "prebuild": "npm run prettier",
     "build": "webpack"
   },
@@ -37,6 +37,7 @@
     "eslint-loader": "^4.0.2",
     "eslint-plugin-react": "^7.20.6",
     "glob": "^7.1.6",
+    "jest": "^26.6.3",
     "mini-css-extract-plugin": "^0.10.0",
     "node-sass": "^4.14.1",
     "nodemon": "^2.0.4",
