+++ src/frontend/app.js
@@ -1,4 +1,4 @@
-import '../assets/styles/main.scss';
+import './assets/styles/main.scss';
 
 import React from 'react';
 import ReactDOM from 'react-dom';
