+++ src/frontend/assets/styles/_var.scss
