+++ src/frontend/assets/styles/main.scss
@@ -1,8 +1,8 @@
-@import './views/shop';
-@import './views/categoriesTree';
-@import './views/productList';
-@import './views/cart';
-@import './views/pagination';
+@import 'views/shop';
+@import 'views/categoriesTree';
+@import 'views/productList';
+@import 'views/cart';
+@import 'views/pagination';
 
 .app-root {
   display: grid;
