+++ src/frontend/assets/styles/views/cart.scss
