+++ src/frontend/assets/styles/views/categoriesTree.scss
