+++ src/frontend/assets/styles/views/pagination.scss
