+++ src/frontend/assets/styles/views/productList.scss
