+++ src/frontend/assets/styles/views/shop.scss
