+++ src/frontend/components/pages/home.js
