+++ src/frontend/components/pages/newProduct.js
