+++ src/frontend/components/pages/shop.js
