+++ src/frontend/components/views/categoriesTree.js
