+++ src/frontend/components/views/footer.js
