+++ src/frontend/components/views/main.js
