+++ src/frontend/components/views/productDetails.js
