+++ src/middleware/routes/api-products.js
